+++ ksa_compliance/patches/_2024_03_20_update_blank_integration_status_in_additional_field.py
@@ -5,5 +5,8 @@ def execute():
     no_of_docs = len(frappe.get_all("Sales Invoice Additional Fields", {"integration_status": ""}, pluck="name"))
     print(f"Update {no_of_docs} Sales Invoice Additional Fields integration status field.")
     frappe.db.sql("""
-        UPDATE `tabSales Invoice Additional Fields` SET integration_status = 'Resend' WHERE integration_status = ''
+        UPDATE 
+            `tabSales Invoice Additional Fields` SET integration_status = 'Resend', docstatus = 0 
+        WHERE 
+            integration_status = ''
     """)